LayoutTests/css3/flexbox/min-size-auto.html

<!DOCTYPE html>
<link href="resources/flexbox.css" rel="stylesheet">

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>

<style>
.flexbox {
    width: 10px;
    height: 10px;
    background-color: grey;
}

.item {
    background-color: red;
    margin: 5px;
}

.child {
    height: 100px;
    width: 100px;
    background-color: green;
}

.width-20 {
    width: 20px;
}

.flex-basis-20 {
    flex-basis: 20px;
}

</style>

<body onload="checkLayout('.flexbox, .inline-flexbox');">
<div id=log></div>

<div class="flexbox" data-expected-width="10">
    <div class="item" data-expected-width="100">
        <div class="child" data-expected-width="100"></div>
    </div>
</div>

<div class="flexbox column" data-expected-height="10">
    <div class="item" data-expected-height="100">
        <div class="child" data-expected-height="100"></div>
    </div>
</div>

<div class="inline-flexbox column" data-expected-height="110">
    <div class="item flex-basis-20" data-expected-height="100">
        <div class="child" data-expected-height="100"></div>
    </div>
</div>

<div class="flexbox" data-expected-width="10">
    <div class="item width-20" data-expected-width="20">
        <div class="child" data-expected-width="100"></div>
    </div>
</div>

<div class="flexbox" data-expected-width="10">
    <div class="item flex-basis-20" data-expected-width="100">
        <div class="child" data-expected-width="100"></div>
    </div>
</div>

<div class="inline-flexbox" data-expected-width="110">
    <div class="item flex-basis-20" data-expected-width="100">
        <div class="child" data-expected-width="100"></div>
    </div>
</div>
